requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
deprecation==2.1.0
distlib==0.3.6
Django==4.1.6
filelock==3.11.0
form==0.0.1
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
lxml==4.9.2
multidict==6.0.4
packaging==23.1
Pillow==9.5.0
platformdirs==3.2.0
PyOpenGL==3.1.6
reportlab==3.6.13
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tqdm==4.64.1
tzdata==2023.3
urllib3==1.26.14
virtualenv==20.21.0
yarl==1.8.2